requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.2
gunicorn==21.1.0
docker==7.1.0
